Add offset = 2 to price notice explanation (#205)
* Escape single-quotes in pop-up macro to fix JS error
Fixes https://github.com/magfest/magprime/issues/200. When we converted our Django templates to Jinja2 templates, we took out a function from the popup macro that turned `'`s into `&quot;`s. We weren't sure if we needed it, but it turns out we do -- it's the only way to properly escape quotes when including the macro inside a quoted element, as in https://github.com/magfest/magprime/issues/200. We also needed to switch the `"`s to '`'s in our magprime template to complete the fix.

* Swap quotes around for better HTML standards