Add estimate for remaining # of badges at a certain price (#213)
* Add estimate for remaining # of badges at a certain price
Fixes https://github.com/magfest/ubersystem/issues/2591. We add the property to the core system because it can be useful, but it's a pretty specific use-case so we only print it in SuperMAG's registration template.

* Define range for estimated badges left text in SuperMAG
We now define what ranges we use inside the template where we use them, avoiding the need to hardcore them anywhere else.

* Small updates to wording for badges left
Renames some of the tiers, and prints the old "Prices may increase before the set date" text if the current # of badges left is unknown.