Update merch for Super 2020 (#294)
* Update merch for Super 2020
Emergency changes to update merch to be the correct prices and descriptions, since merch appears on staff badge pages.

* Re-add sweatpants column
My current dev environment is hosed, migrations-wise, so I can't create new ones. For the time being, it's fine if we keep this column but just not require staffers to fill it out anymore (already done in the previous commit).